Replace getBalance method with utility function
